Recognize the new name for a study paper.
This was introduced in the new [examination regulations of 2022](https://www.uni-stuttgart.de/universitaet/aktuelles/bekanntmachungen/dokumente/bekanntm_30_2022.pdf)
The term "Forschungsarbeit" is the new default for master's students, so it will be the dominant term, soon.
This commit keeps the old term "Studienarbeit" with the key "study-old", as there are still students working under pre-2022 regulations.